+++ ImageFeed/ImagesList/ImagesListService.swift
@@ -1,10 +1,3 @@
-//
-//  ImagesListService.swift
-//  ImageFeed
-//
-//  Created by Кирилл Брызгунов on 26.02.2023.
-//
-
 import Foundation
 import SwiftKeychainWrapper
 
@@ -13,11 +6,6 @@ class ImagesListService {
     static let didChangeNotification = Notification.Name(rawValue: "ImageListServiceDidChange")
     private let urlSession = URLSession.shared
     static let shared = ImagesListService()
-    //private let httpClient: HttpClient
-    
-//    init(httpClient: HttpClient) {
-//        self.httpClient = httpClient
-//    }
     
     private lazy var dateFormatter: ISO8601DateFormatter = {
         return ISO8601DateFormatter()
@@ -35,8 +23,6 @@ class ImagesListService {
     private var task: URLSessionTask?
     private let tokenStorage = SwiftKeychainWrapper()
     
-    //private(set) var profile: Profile?
-    
     func fetchPhotosNextPage() {
         let nextPage: Page
         if let lastLoadedPage = lastLoadedPage {
@@ -46,22 +32,17 @@ class ImagesListService {
             nextPage = Page(number: 1, perPage: 10)
         }
 
-        // Already fetching for a given authToken
         guard task == nil || self.loadingPage != nextPage else { return }
-        // In case a task for a different token is in progress - cancel it
         task?.cancel()
-        // Remember the page to prevent double-loading
         self.loadingPage = nextPage
 
         let request = photosRequest(page: nextPage.number, perPage: nextPage.perPage)
-        //let task = urlSession.objectTask(for: request) { (result: Result<ProfileResult, Error>) in
         let task = urlSession.objectTask(for: request, completion: { [weak self] (result: Result<[PhotoResult], Error>) in
             guard let self = self else { return }
 
             self.task = nil
 
             guard case .success(let photosResult) = result else {
-                // Don't do anything, another fetch will be triggered if user scrolls down
                 return
             }
 
@@ -85,11 +66,6 @@ class ImagesListService {
         task.resume()
         self.task = task
     }
-
-    
-//    private func convert(photoResult: PhotoResult) -> Photo {
-//        return Photo(from: photoResult)
-//    }
     
     private func photosRequest(page: Int, perPage: Int) -> URLRequest {
         makeRequest(path: "/photos?page=\(page)&&per_page=\(perPage)")
@@ -115,27 +91,6 @@ class ImagesListService {
 }
 
 extension ImagesListService {
-//    func changeLike(photoId: String, isLike: Bool, _ completion: @escaping (Result<Void, Error>) -> Void) {
-//        let request = isLike ? likeRequest(photoId: photoId) : unlikeRequest(photoId: photoId)
-//
-//        let task = urlSession.objectTask(for: request, completion: { [weak self] (result: Result<[PhotoResult], Error>) in
-//            //let task = httpsClient.data(for: request) { result in
-//            switch result {
-//            case .success:
-//                guard let updatedPhotoIndex = self!.photos.firstIndex(where: { $0.id == photoId })
-//                else { return }
-//                let oldPhoto = self!.photos[updatedPhotoIndex]
-//                let newPhoto = oldPhoto.withToggledLike()
-//                self!.photos = self!.photos.withReplaced(itemAt: updatedPhotoIndex, newValue: newPhoto)
-//                completion(.success(()))
-//
-//            case .failure(let error):
-//                completion(.failure(error))
-//            }
-//        }
-//        task.resume()
-//    }
-    
     func changeLike(photoId: String, isLike: Bool, _ completion: @escaping (Result<Void, Error>) -> Void) {
          enum LikeError: Error {
              case photoNotFound
@@ -186,68 +141,3 @@ extension ImagesListService {
     }
 }
 
-struct Photo {
-    let id: String
-    let size: CGSize
-    let createdAt: Date?
-    let welcomeDescription: String?
-    let thumbImageURL: String
-    let largeImageURL: String
-    let isLiked: Bool
-}
-
-extension Photo {
-    func withToggledLike() -> Photo {
-        Photo(
-            id: id,
-            size: size,
-            createdAt: createdAt,
-            welcomeDescription: welcomeDescription,
-            thumbImageURL: thumbImageURL,
-            largeImageURL: largeImageURL,
-            isLiked: !isLiked
-        )
-    }
-}
-
-struct Page: Equatable {
-    let number: Int
-    let perPage: Int
-}
-
-struct UrlsResult: Codable {
-    let raw, full, regular, small: String
-    let thumb: String
-}
-
-extension Array {
-    func withReplaced(itemAt index: Int, newValue: Element) -> Self {
-        var result = self
-        result[index] = newValue
-        return result
-    }
-}
-
-struct LikeResult: Decodable {
-    let photo: PhotoResult
-}
-
-struct PhotoResult: Decodable {
-    let id: String
-    let createdAt, updatedAt: String
-    let width, height: Int
-    let color, blurHash, welcomeDescription: String?
-    let urls: UrlsResult
-    let likedByUser: Bool
-
-    enum CodingKeys: String, CodingKey {
-        case id
-        case createdAt = "created_at"
-        case updatedAt = "updated_at"
-        case width, height, color
-        case blurHash = "blur_hash"
-        case welcomeDescription = "description"
-        case urls
-        case likedByUser = "liked_by_user"
-    }
-}
