--- ImageFeed/ImagesList/PhotoResult .swift
@@ -1,62 +1,0 @@
-//
-//  PhotoResult .swift
-//  ImageFeed
-//
-//  Created by Кирилл Брызгунов on 26.02.2023.
-//
-
-
-import Foundation
-
-/*
-// Структуры
-struct PhotoResult: Decodable {
-    let id: String
-    let createdAt: String
-    let updatedAt: String
-    let width: Int
-    let height: Int
-    let color: String
-    let blurHash: String
-    let likes: Int
-    let likedByUser: Bool
-    let description: String
-    //let user: ProfileResult
-    let urls: UrlsResult
-    
-    enum CodingKeys: String, CodingKey {
-        case id = "id"
-        case createdAt = "created_at"
-        case updatedAt = "updated_at"
-        case width = "width"
-        case height = "height"
-        case color = "color"
-        case blurHash = "blur_hash"
-        case likes = "likes"
-        case likedByUser = "liked_by_user"
-        case description = "description"
-        //case user = "user"
-        case urls = "urls"
-    }
-}
-
-*/
-
-
-
-//struct UrlsResult: Decodable {
-//    let raw: String
-//    let full: String
-//    let regular: String
-//    let small: String
-//    let thumb: String
-//
-//    enum CodingKeys: String, CodingKey {
-//        case raw = "raw"
-//        case full = "full"
-//        case regular = "regular"
-//        case small = "small"
-//        case thumb = "thumb"
-//    }
-//}
-
