+++ ImageFeed/Network/HttpClient.swift
@@ -3,18 +3,6 @@ import Foundation
 
 // MARK: - HTTP Requset
 
-//extension URLRequest {
-//    static func makeHTTPRequest(
-//        path: String,
-//        httpMethod: String,
-//        baseURL: URL = defaultBaseURL
-//    ) -> URLRequest {
-//        var request = URLRequest(url: URL(string: path, relativeTo: baseURL)!)
-//        request.httpMethod = httpMethod
-//        return request
-//    }
-//}
-
 enum HttpClientError: Error {
     case httpStatusCode(Int)
     case urlRequestError(Error)
