+++ ImageFeed/Network/OAuth2Service.swift
@@ -1,10 +1,3 @@
-//
-//  OAuth2Service.swift
-//  ImageFeed
-//
-//  Created by Кирилл Брызгунов on 08.02.2023.
-//
-
 import Foundation
 import SwiftKeychainWrapper
 
@@ -58,19 +51,7 @@ final class OAuth2Service {
 }
 
 extension OAuth2Service {
-//    private func object(
-//        for request: URLRequest,
-//        completion: @escaping (Result<OAuthTokenResponseBody, Error>) -> Void
-//    ) -> URLSessionTask {
-//        let decoder = JSONDecoder()
-//        return urlSession.data(for: request) { (result: Result<Data, Error>) in
-//            let response = result.flatMap { data -> Result<OAuthTokenResponseBody, Error> in
-//                Result { try decoder.decode(OAuthTokenResponseBody.self, from: data) }
-//            }
-//            completion(response)
-//        }
-//    }
-
+    
     private func authTokenRequest(code: String) -> URLRequest? {
         let urlString = "https://unsplash.com/oauth/token"
         var urlComponents = URLComponents(string: urlString)
