+++ ImageFeed/Network/SwiftKeychainWrapper .swift
@@ -1,10 +1,3 @@
-//
-//  SwiftKeychainWrapper .swift
-//  ImageFeed
-//
-//  Created by Кирилл Брызгунов on 24.02.2023.
-//
-
 import Foundation
 import SwiftKeychainWrapper
