+++ ImageFeed/ProfileView/ProfileService.swift
@@ -51,7 +51,7 @@ final class ProfileService {
         
         var request = URLRequest(url: url)
         request.httpMethod = "GET"
-        request.setValue("Bearer \(token)", forHTTPHeaderField: "Authorization") // не до конца понятна
+        request.setValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
         
         return request
     }
