+++ ImageFeed/SingleImageView/SingleImageViewController.swift
@@ -3,14 +3,6 @@ import Kingfisher
 import UIKit
 
 final class SingleImageViewController: UIViewController {
-//    var image: UIImage! {
-//        didSet {
-//            guard isViewLoaded else { return }
-//            imageView.image = image
-//            rescaleAndCenterImageInScrollView(image: image)
-//        }
-//    }
-    
     var largeImageUrl: URL?
     
     @IBOutlet var imageView: UIImageView!
@@ -124,37 +116,4 @@ extension SingleImageViewController {
         present(alertController, animated: true)
     }
 }
-
-
-//override func viewDidLoad() {
-//    super.viewDidLoad()
-//    imageView.image = image
-//    rescaleAndCenterImageInScrollView(image: image)
-//
-//    scrollView.minimumZoomScale = 0.1
-//    scrollView.maximumZoomScale = 1.25
-//}
-//
-//private func rescaleAndCenterImageInScrollView(image: UIImage) {
-//    let minZoomScale = scrollView.minimumZoomScale
-//    let maxZoomScale = scrollView.maximumZoomScale
-//    view.layoutIfNeeded()
-//    let visibleRectSize = scrollView.bounds.size
-//    let imageSize = image.size
-//    let hScale = visibleRectSize.width / imageSize.width
-//    let vScale = visibleRectSize.height / imageSize.height
-//    let scale = min(maxZoomScale, max(minZoomScale, max(hScale, vScale)))
-//    scrollView.setZoomScale(scale, animated: false)
-//    scrollView.layoutIfNeeded()
-//    let newContentSize = scrollView.contentSize
-//    let x = (newContentSize.width - visibleRectSize.width) / 2
-//    let y = (newContentSize.height - visibleRectSize.height) / 2
-//    scrollView.setContentOffset(CGPoint(x: x, y: y), animated: false)
-//}
-//}
-//
-//extension SingleImageViewController: UIScrollViewDelegate {
-//func viewForZooming(in scrollView: UIScrollView) -> UIView? {
-//    imageView
-//}
